+++ .eslintrc.js
@@ -47,6 +47,7 @@ module.exports = {
     'newline-per-chained-call': 'off',
     'no-console': 'error',
     'no-const-assign': 'error',
+    'no-continue': 'off',
     'no-else-return': 'off',
     'no-labels': 'off',
     'no-mixed-operators': 'off',