+++ packages/shared-metrics/test/.eslintrc.js
@@ -12,7 +12,12 @@ module.exports = {
   root: true,
 
   env: {
+    es6: true,
     mocha: true
+  },
+
+  parserOptions: {
+    ecmaVersion: 2018
   },
 
   rules: {